requirements.txt
litellm>=1.0.0
tenacity>=8.0.0
pyyaml>=6.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
